+++ frontend/src/board/Board.tsx
@@ -9,23 +9,42 @@ import { gql, useQuery } from "@apollo/client"
 import { RouteComponentProps, withRouter } from "react-router-dom"
 import TrickCallDialog from "./TrickCallDialog"
 import { withSnackbar, WithSnackbarProps } from "notistack"
+import useSound from "use-sound"
+import { useSettings } from "../settings"
 
 interface Props extends WithStyles<typeof styles>, RouteComponentProps, WithSnackbarProps { }
 
 function Board(props: Props) {
+	const { settings } = useSettings()
 	const { data: gameInfo } = useQuery<GameInfo>(GAME_INFO, { pollInterval: 1000 })
+	const [playTurnSound] = useSound("/your_turn_0.mp3")
 
 	React.useEffect(() => {
-		if (gameInfo?.playerTurn === localStorage.getItem("name")) {
-			props.enqueueSnackbar("Your turn.", {
-				autoHideDuration: null,
-				variant: "info",
-				key: "player-turn"
-			})
-		} else {
-			props.closeSnackbar("player-turn")
+		if (settings.notifications.enabled) {
+			if (gameInfo?.playerTurn === localStorage.getItem("name")) {
+				if (settings.notifications.desktop) {
+					Notification.requestPermission()
+						.then(() => {
+							new Notification("Wizard", {
+								body: "Du bist am Zug.",
+								silent: !settings.notifications.audio
+							})
+						})
+				} else {
+					props.enqueueSnackbar("Your turn.", {
+						autoHideDuration: null,
+						variant: "info",
+						key: "player-turn"
+					})
+					if (settings.notifications.audio) {
+						playTurnSound()
+					}
+				}
+			} else if (!settings.notifications.desktop) {
+				props.closeSnackbar("player-turn")
+			}
 		}
-	}, [gameInfo?.playerTurn, props])
+	}, [gameInfo?.playerTurn, props, playTurnSound, settings])
 
 	if (gameInfo?.gameInProgress === false) {
 		props.history.push("/")