+++ frontend/src/settings/index.ts
@@ -1,0 +1,3 @@
+export { default as SettingsProvider, useSettings } from "./SettingsProvider"
+export { default as SettingsDialog } from "./SettingsDialog"
+export type { default as Settings } from "./type"